feat(acw897): add solution for the problem ACW897 in Go
This commit adds a solution for problem ACW897 in Go. The solution calculates the length of the longest common subsequence between two strings, `aa` and `bb`. It reads the input from a file named `testcases.txt` and prints the result.

The solution uses a dynamic programming approach to calculate the length of the longest common subsequence. It initializes a 2D array `f` of size `(n+1) x (m+1)`, where `n` and `m` are the lengths of `aa` and `bb` respectively. The value `f[i][j]` represents the length of the longest common subsequence between the substrings `aa[0,...,i-1]` and `bb[0,...,j-1]`.

The algorithm iterates over the substrings `aa` and `bb` and fills the `f` array based on the following rules:
- If `aa[i-1]` is equal to `bb[j-1]`, then `f[i][j]` is set to `f[i-1][j-1] + 1`.
- Otherwise, `f[i][j]` is set to the maximum value between `f[i-1][j]` and `f[i][j-1]`.

Finally, the solution prints the value of `f[n][m]`, which represents the length of the longest common subsequence between `aa` and `bb`.

The commit also includes a new file named `testcases.txt` which contains the test input for the problem.