feat(acw278): add solution for ACW278 problem in Go
This commit adds a solution for the ACW278 problem in Go. The solution reads input from a file called "testcases.txt" and performs calculations based on the given input.

The "acw278.go" file contains the main code for the solution. It opens the "testcases.txt" file, reads the values of n and m, and then reads an array of integers. It uses dynamic programming to calculate the number of ways to select a subset of the array elements that sum up to m. The result is printed to the console.

A new file called "testcases.txt" is also added, which contains the test case used in the solution.